WIP padding, passing signature checks
